# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TAOX API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/providers": {
            "get": {
                "tags": [
                    "providers"
                ],
                "summary": "Get Providers",
                "description": "Obtiene todos los proveedores de una empresa",
                "operationId": "get_providers_providers_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProviderDTO"
                                    },
                                    "title": "Response Get Providers Providers Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "providers"
                ],
                "summary": "Create Provider",
                "description": "Crea un nuevo proveedor",
                "operationId": "create_provider_providers_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProviderCreateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProviderDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/providers/{provider_id}": {
            "get": {
                "tags": [
                    "providers"
                ],
                "summary": "Get Provider",
                "description": "Obtiene un proveedor específico",
                "operationId": "get_provider_providers__provider_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "provider_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Provider Id"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProviderDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "providers"
                ],
                "summary": "Update Provider",
                "description": "Actualiza un proveedor",
                "operationId": "update_provider_providers__provider_id__put",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "provider_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Provider Id"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Company Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProviderUpdateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProviderDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "providers"
                ],
                "summary": "Delete Provider",
                "description": "Elimina un proveedor",
                "operationId": "delete_provider_providers__provider_id__delete",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "provider_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Provider Id"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ProviderCreateDTO": {
                "properties": {
                    "cad_proveedor": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1,
                        "title": "Cad Proveedor",
                        "description": "Código único del proveedor"
                    },
                    "nombre": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1,
                        "title": "Nombre",
                        "description": "Nombre del proveedor"
                    },
                    "contacto": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1,
                        "title": "Contacto",
                        "description": "Persona de contacto"
                    },
                    "direccion": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1,
                        "title": "Direccion",
                        "description": "Dirección"
                    },
                    "telefono": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 7,
                        "title": "Telefono",
                        "description": "Teléfono"
                    },
                    "celular": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 7,
                        "title": "Celular",
                        "description": "Celular"
                    },
                    "web": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Web",
                        "description": "Sitio web"
                    },
                    "correo": {
                        "type": "string",
                        "title": "Correo",
                        "description": "Email del proveedor"
                    },
                    "company_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Company Id",
                        "description": "UUID de la empresa"
                    }
                },
                "type": "object",
                "required": [
                    "cad_proveedor",
                    "nombre",
                    "contacto",
                    "direccion",
                    "telefono",
                    "celular",
                    "correo",
                    "company_id"
                ],
                "title": "ProviderCreateDTO",
                "description": "DTO para crear un proveedor"
            },
            "ProviderDTO": {
                "properties": {
                    "cad_proveedor": {
                        "type": "string",
                        "title": "Cad Proveedor"
                    },
                    "nombre": {
                        "type": "string",
                        "title": "Nombre"
                    },
                    "contacto": {
                        "type": "string",
                        "title": "Contacto"
                    },
                    "direccion": {
                        "type": "string",
                        "title": "Direccion"
                    },
                    "telefono": {
                        "type": "string",
                        "title": "Telefono"
                    },
                    "celular": {
                        "type": "string",
                        "title": "Celular"
                    },
                    "web": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Web"
                    },
                    "correo": {
                        "type": "string",
                        "title": "Correo"
                    },
                    "company_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Company Id"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    }
                },
                "type": "object",
                "required": [
                    "cad_proveedor",
                    "nombre",
                    "contacto",
                    "direccion",
                    "telefono",
                    "celular",
                    "web",
                    "correo",
                    "company_id",
                    "is_active"
                ],
                "title": "ProviderDTO",
                "description": "DTO para respuesta de proveedor"
            },
            "ProviderUpdateDTO": {
                "properties": {
                    "nombre": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 100,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Nombre"
                    },
                    "contacto": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 100,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Contacto"
                    },
                    "direccion": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Direccion"
                    },
                    "telefono": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 20,
                                "minLength": 7
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Telefono"
                    },
                    "celular": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 20,
                                "minLength": 7
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Celular"
                    },
                    "web": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Web"
                    },
                    "correo": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Correo"
                    },
                    "is_active": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Is Active"
                    }
                },
                "type": "object",
                "title": "ProviderUpdateDTO",
                "description": "DTO para actualizar un proveedor"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}